refactor(backend): extract AgentRoster from OpenBotDatabase
Moves `listAgents`, `replaceAgents` and `hardDeleteAgent` into
`src/backend/database/agent-roster.ts`. `#ensureThreadProjection` moves with
them and becomes public on the controller — the conversation writer calls it
from inside its own dispatch and already passes the `db` that dispatch gave it —
but is deliberately not exposed on the facade, whose surface stays frozen.

`hardDeleteAgent` is moved statement for statement. Every receipt DELETE selects
its command ids out of `orchestration_events`, so receipts must be deleted
before the matching events in all four blocks; reordering them would leave
orphan receipts that `dispatch` later replays as a stale result for a fresh
command.
